Vaccine Distribution System
Our project is about Vaccine Distribution System. The working of our project is as follows:

These events are not dependent on each other –

Initially, the local Health Department Admin adds the population of a created state.
The manufacturer manufactures a particular drug.
The CDC(Centers for Disease Control and Prevention) officer places an order for a drug to the manufacturer.
Providers (hospital provider/remote clinic provider) place an order for a drug that they need.

The events mentioned below happen in sequence –

First, the manufacturer ships the drug ordered by the CDC to a distributor.
The distributor then stores this drug in a warehouse.
If the remote clinic provider has placed a request for a particular drug, this order first needs to be approved by the hospital provider. Then it follows the same sequence as below.
The local Health Department admin approves the order placed by the providers.
This order then needs to be approved by the CDC officer.
Once the order is approved by the CDC officer, the distributor is permitted to ship the same to the providers.
The providers add the received drugs to their inventory.
The providers can then administer this drug to the patients.